master 서버 포트 변경

See merge request s09-webmobile1-sub2/S09P12E202!74

+++ backend/src/main/java/com/ssafish/domain/deck/DeckRepository.java
@@ -1,9 +1,8 @@
 package com.ssafish.domain.deck;
 
-import com.ssafish.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeckRepository extends JpaRepository<Deck, Integer> {
+public interface DeckRepository extends JpaRepository<Deck,Long> {
